Backend/requirements.txt
fastapi
playwright
pydantic
python-dotenv
supabase